added separate components for editablecell & editablerow
