+++ src/components/BulkEditAccessPoints/components/BulkEditAPTable/index.module.scss
@@ -2,19 +2,20 @@
   position: relative;
 }
 
-.editableCellValueWrap {
-  padding: 5px 12px;
-  cursor: pointer;
+.editableRow {
+  .ant-table-cell {
+    &:focus {
+      outline: none;
+    }
+  }
 }
 
-.editableRow:hover .editableCellValueWrap {
-  border: 1px solid #d9d9d9;
-  border-radius: 4px;
-  padding: 4px 11px;
-}
-
-[data-theme='dark'] .editableRow:hover .editableCellValueWrap {
-  border: 1px solid #434343;
+[data-theme='dark'] {
+  .editableRow {
+    &:hover {
+      border-color: #434343;
+    }
+  }
 }
 
 .LoadMore {
