+++ src/components/LocationsTree/components/PopoverMenuContent/index.js
@@ -70,11 +70,12 @@ PopoverMenuContent.propTypes = {
   setDeleteModal: PropTypes.func,
   setBulkEditApIds: PropTypes.func,
   locationType: PropTypes.string,
-  locationId: PropTypes.number.isRequired,
+  locationId: PropTypes.number,
 };
 
 PopoverMenuContent.defaultProps = {
   locationType: null,
+  locationId: 0,
   setEditModal: () => {},
   setAddModal: () => {},
   setDeleteModal: () => {},
